+++ backend/models/mongo/db_init.py
@@ -5,7 +5,8 @@ This file contains the initialization code for MongoDB connection using Beanie.
 from motor.motor_asyncio import AsyncIOMotorClient
 from beanie import init_beanie
 
-from backend.models.mongodb import WebContent, MongoDBManager
+from backend.models.mongo.web_content import WebContent
+from backend.models.mongo.db_manager import MongoDBManager
 
 
 async def init_mongodb(mongodb_url: str, database_name: str):
